+++ src/emulator/functionsEmulator.ts
@@ -5,6 +5,7 @@ import * as clc from "colorette";
 import * as http from "http";
 import * as jwt from "jsonwebtoken";
 import * as cors from "cors";
+import * as semver from "semver";
 import { URL } from "url";
 import { EventEmitter } from "events";
 
@@ -85,8 +86,8 @@ export interface EmulatableBackend {
   secretEnv: backend.SecretEnvVar[];
   codebase: string;
   predefinedTriggers?: ParsedTriggerDefinition[];
-  nodeMajorVersion?: number;
-  nodeBinary?: string;
+  runtime?: string;
+  bin?: string;
   extensionInstanceId?: string;
   extension?: Extension; // Only present for published extensions
   extensionVersion?: ExtensionVersion; // Only present for published extensions
@@ -336,7 +337,12 @@ export class FunctionsEmulator implements EmulatorInstance {
     const record = this.getTriggerRecordByKey(this.getTriggerKey(trigger));
     const pool = this.workerPools[record.backend.codebase];
     if (!pool.readyForWork(trigger.id)) {
-      await this.startRuntime(record.backend, trigger);
+      try {
+        await this.startRuntime(record.backend, trigger);
+      } catch (e: any) {
+        this.logger.logLabeled("ERROR", `Failed to start runtime for ${trigger.id}: ${e}`);
+        return;
+      }
     }
     const worker = pool.getIdleWorker(trigger.id)!;
     const reqBody = JSON.stringify(body);
@@ -360,9 +366,6 @@ export class FunctionsEmulator implements EmulatorInstance {
   }
 
   async start(): Promise<void> {
-    for (const backend of this.args.emulatableBackends) {
-      backend.nodeBinary = this.getNodeBinary(backend);
-    }
     const credentialEnv = await this.getCredentialsEnvironment();
     for (const e of this.args.emulatableBackends) {
       e.env = { ...credentialEnv, ...e.env };
@@ -452,15 +455,17 @@ export class FunctionsEmulator implements EmulatorInstance {
         projectId: this.args.projectId,
         projectDir: this.args.projectDir,
         sourceDir: emulatableBackend.functionsDir,
+        runtime: emulatableBackend.runtime,
       };
-      if (emulatableBackend.nodeMajorVersion) {
-        runtimeDelegateContext.runtime = `nodejs${emulatableBackend.nodeMajorVersion}`;
-      }
       const runtimeDelegate = await runtimes.getRuntimeDelegate(runtimeDelegateContext);
       logger.debug(`Validating ${runtimeDelegate.name} source`);
       await runtimeDelegate.validate();
       logger.debug(`Building ${runtimeDelegate.name} source`);
       await runtimeDelegate.build();
+
+      // Retrieve information from the runtime delegate.
+      emulatableBackend.runtime = runtimeDelegate.runtime;
+      emulatableBackend.bin = runtimeDelegate.bin;
 
       // Don't include user envs when parsing triggers. Do include user envs when resolving parameter values
       const firebaseConfig = this.getFirebaseConfig();
@@ -499,12 +504,6 @@ export class FunctionsEmulator implements EmulatorInstance {
    * TODO(b/216167890): Gracefully handle removal of deleted function definitions
    */
   async loadTriggers(emulatableBackend: EmulatableBackend, force = false): Promise<void> {
-    if (!emulatableBackend.nodeBinary) {
-      throw new FirebaseError(
-        `No node binary for ${emulatableBackend.functionsDir}. This should never happen.`
-      );
-    }
-
     let triggerDefinitions: EmulatedTriggerDefinition[] = [];
     try {
       triggerDefinitions = await this.discoverTriggers(emulatableBackend);
@@ -674,7 +673,14 @@ export class FunctionsEmulator implements EmulatorInstance {
           {}
         )
       );
-      await this.startRuntime(emulatableBackend);
+      try {
+        await this.startRuntime(emulatableBackend);
+      } catch (e: any) {
+        this.logger.logLabeled(
+          "ERROR",
+          `Failed to start functions in ${emulatableBackend.functionsDir}: ${e}`
+        );
+      }
     }
   }
 
@@ -1029,72 +1035,6 @@ export class FunctionsEmulator implements EmulatorInstance {
     triggers.forEach((def) => this.addTriggerRecord(def, { backend, ignored: false }));
   }
 
-  getNodeBinary(backend: EmulatableBackend): string {
-    const pkg = require(path.join(backend.functionsDir, "package.json"));
-    // If the developer hasn't specified a Node to use, inform them that it's an option and use default
-    if ((!pkg.engines || !pkg.engines.node) && !backend.nodeMajorVersion) {
-      this.logger.log(
-        "WARN",
-        `Your functions directory ${backend.functionsDir} does not specify a Node version.\n   ` +
-          "- Learn more at https://firebase.google.com/docs/functions/manage-functions#set_runtime_options"
-      );
-      return process.execPath;
-    }
-
-    const hostMajorVersion = process.versions.node.split(".")[0];
-    const requestedMajorVersion: string = backend.nodeMajorVersion
-      ? `${backend.nodeMajorVersion}`
-      : pkg.engines.node;
-    let localMajorVersion = "0";
-    const localNodePath = path.join(backend.functionsDir, "node_modules/.bin/node");
-
-    // Next check if we have a Node install in the node_modules folder
-    try {
-      const localNodeOutput = spawn.sync(localNodePath, ["--version"]).stdout.toString();
-      localMajorVersion = localNodeOutput.slice(1).split(".")[0];
-    } catch (err: any) {
-      // Will happen if we haven't asked about local version yet
-    }
-
-    // If the requested version is already locally available, let's use that
-    if (requestedMajorVersion === localMajorVersion) {
-      this.logger.logLabeled(
-        "SUCCESS",
-        "functions",
-        `Using node@${requestedMajorVersion} from local cache.`
-      );
-      return localNodePath;
-    }
-
-    // If the requested version is the same as the host, let's use that
-    if (requestedMajorVersion === hostMajorVersion) {
-      this.logger.logLabeled(
-        "SUCCESS",
-        "functions",
-        `Using node@${requestedMajorVersion} from host.`
-      );
-    } else {
-      // Otherwise we'll warn and use the version that is currently running this process.
-      if (process.env.FIREPIT_VERSION) {
-        this.logger.log(
-          "WARN",
-          `You've requested "node" version "${requestedMajorVersion}", but the standalone Firebase CLI comes with bundled Node "${hostMajorVersion}".`
-        );
-        this.logger.log(
-          "INFO",
-          `To use a different Node.js version, consider removing the standalone Firebase CLI and switching to "firebase-tools" on npm.`
-        );
-      } else {
-        this.logger.log(
-          "WARN",
-          `Your requested "node" version "${requestedMajorVersion}" doesn't match your global version "${hostMajorVersion}". Using node@${hostMajorVersion} from host.`
-        );
-      }
-    }
-
-    return process.execPath;
-  }
-
   getRuntimeConfig(backend: EmulatableBackend): Record<string, string> {
     const configPath = `${backend.functionsDir}/.runtimeconfig.json`;
     try {
@@ -1268,18 +1208,18 @@ export class FunctionsEmulator implements EmulatorInstance {
     return secretEnvs;
   }
 
-  async startRuntime(
+  async startNode(
     backend: EmulatableBackend,
-    trigger?: EmulatedTriggerDefinition
-  ): Promise<RuntimeWorker> {
-    const emitter = new EventEmitter();
+    envs: Record<string, string>
+  ): Promise<FunctionsRuntimeInstance> {
     const args = [path.join(__dirname, "functionsEmulatorRuntime")];
-
     if (this.args.debugPort) {
-      if (process.env.FIREPIT_VERSION && process.execPath === backend.nodeBinary) {
+      if (process.env.FIREPIT_VERSION) {
         this.logger.log(
           "WARN",
-          `To enable function inspection, please run "${process.execPath} is:npm i node@${backend.nodeMajorVersion} --save-dev" in your functions directory`
+          `To enable function inspection, please run "npm i node@${semver.coerce(
+            backend.runtime || "18.0.0"
+          )} --save-dev" in your functions directory`
         );
       } else {
         const { host } = this.getInfo();
@@ -1301,28 +1241,43 @@ export class FunctionsEmulator implements EmulatorInstance {
           "See https://yarnpkg.com/getting-started/migration#step-by-step for more information."
       );
     }
-    const runtimeEnv = this.getRuntimeEnvs(backend, trigger);
-    const secretEnvs = await this.resolveSecretEnvs(backend, trigger);
-    const socketPath = getTemporarySocketPath();
 
-    const childProcess = spawn(backend.nodeBinary!, args, {
+    const bin = backend.bin;
+    if (!bin) {
+      throw new Error(
+        `No binary associated with ${backend.functionsDir}. ` +
+          "Make sure function runtime is configured correctly in firebase.json."
+      );
+    }
+
+    const socketPath = getTemporarySocketPath();
+    const childProcess = spawn(bin, args, {
       cwd: backend.functionsDir,
       env: {
-        node: backend.nodeBinary,
+        node: backend.bin,
         ...process.env,
-        ...runtimeEnv,
-        ...secretEnvs,
+        ...envs,
         PORT: socketPath,
       },
       stdio: ["pipe", "pipe", "pipe", "ipc"],
     });
 
-    const runtime: FunctionsRuntimeInstance = {
+    return Promise.resolve({
       process: childProcess,
-      events: emitter,
+      events: new EventEmitter(),
       cwd: backend.functionsDir,
       socketPath,
-    };
+    });
+  }
+
+  async startRuntime(
+    backend: EmulatableBackend,
+    trigger?: EmulatedTriggerDefinition
+  ): Promise<RuntimeWorker> {
+    const runtimeEnv = this.getRuntimeEnvs(backend, trigger);
+    const secretEnvs = await this.resolveSecretEnvs(backend, trigger);
+
+    const runtime = await this.startNode(backend, { ...runtimeEnv, ...secretEnvs });
     const extensionLogInfo = {
       instanceId: backend.extensionInstanceId,
       ref: backend.extensionVersion?.ref,
@@ -1482,7 +1437,16 @@ export class FunctionsEmulator implements EmulatorInstance {
 
     const pool = this.workerPools[record.backend.codebase];
     if (!pool.readyForWork(trigger.id)) {
-      await this.startRuntime(record.backend, trigger);
+      try {
+        await this.startRuntime(record.backend, trigger);
+      } catch (e: any) {
+        this.logger.logLabeled("ERROR", `Failed to handle request for function ${trigger.id}`);
+        this.logger.logLabeled(
+          "ERROR",
+          `Failed to start functions in ${record.backend.functionsDir}: ${e}`
+        );
+        return;
+      }
     }
     const debugBundle = this.args.debugPort
       ? {