+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-- Fix bug where CLI was unable to deploy Firebase Functions in some monorepo setups (#5391)
-- Upgrade Storage Rules Runtime to v1.1.3 to support ternary operators (#5370)
-- Fixes an issue where already deployed functions with the same remote configuration do not get skipped (#5354)
+- Refactor Functions Emulator. (#5422)
+- Fix race condition when discovering functions. (#5444)